Extract hover/pointer detection to useMediaQuery
