+++ src/components/widgets/compilation/CompilationItemListRow.tsx
@@ -13,25 +13,26 @@ import IconBtn from '@/components/ui/IconBtn'
 import ReadIconBtn from '@/components/ui/ReadIconBtn'
 import Tooltip from '@/components/ui/Tooltip'
 import ConfirmDialog from '@/components/widgets/ConfirmDialog'
-import { EPISODE_ROW_ACTION_BTN_CLASS } from '@/lib/episode'
 import type { SortableListDragHandleProps } from '@/components/widgets/SortableList'
 import { useLibrary } from '@/contexts/LibraryContext'
 import type { PlayerQueueItem } from '@/contexts/MediaContext'
 import { useMediaContext } from '@/contexts/MediaContext'
-import { usePrimaryInputCanHover, useSortableCompilation } from '@/contexts/SortableCompilationContext'
+import { useSortableCompilation } from '@/contexts/SortableCompilationContext'
 import { useGlobalToast } from '@/contexts/ToastContext'
 import { useUser } from '@/contexts/UserContext'
 import { useCompilationListRowLayout } from '@/hooks/useCompilationListRowLayout'
+import { usePrimaryInputCanHover } from '@/hooks/useMediaQuery'
 import { useEpisodeListenActions, type PodcastEpisodeListenActions } from '@/hooks/usetEpisodeListenActions'
 import { useTypeSafeTranslations } from '@/hooks/useTypeSafeTranslations'
 import { getMediaCardModalNavigationContext } from '@/lib/bookshelfNavigationContext'
-import { buildBookQueueItem, buildEpisodeQueueItem } from '@/lib/playerQueue'
 import { getLibraryItemCoverSrc, getPlaceholderCoverUrl } from '@/lib/coverUtils'
 import { DRAG_HANDLE_COARSE_POINTER_MIN_TOUCH, DRAG_HANDLE_GRAB_CURSOR } from '@/lib/dragHandleClasses'
+import { EPISODE_ROW_ACTION_BTN_CLASS } from '@/lib/episode'
 import { getMediaCardEpisodeEditNavigationContext } from '@/lib/episodeEditNavigation'
 import { formatDuration } from '@/lib/formatDuration'
 import { buildMediaItemProgressMap, buildPodcastEpisodeProgressMap, getLibraryItemProgressFromMap } from '@/lib/mediaProgress'
 import { mergeClasses } from '@/lib/merge-classes'
+import { buildBookQueueItem, buildEpisodeQueueItem } from '@/lib/playerQueue'
 import { getPlaylistItemDuration } from '@/lib/playlistItems'
 import type { ShelfNavigationEntity } from '@/lib/shelfNavigationEntity'
 import type { LibraryItem, PodcastEpisode } from '@/types/api'
