+++ src/contexts/SortableCompilationContext.tsx
@@ -1,31 +1,6 @@
 'use client'
 
-import { createContext, useContext, useSyncExternalStore, type ReactNode } from 'react'
-
-const HOVER_MEDIA_QUERY = '(hover: hover)'
-
-function subscribeHoverMedia(onStoreChange: () => void) {
-  const mq = window.matchMedia(HOVER_MEDIA_QUERY)
-  mq.addEventListener('change', onStoreChange)
-  return () => mq.removeEventListener('change', onStoreChange)
-}
-
-function getHoverMediaSnapshot() {
-  return window.matchMedia(HOVER_MEDIA_QUERY).matches
-}
-
-/** SSR / first paint: assume hover-capable to match typical desktop layouts. */
-function getHoverMediaServerSnapshot() {
-  return true
-}
-
-/**
- * True when the primary input can hover (e.g. mouse / trackpad). False on touch-first UIs
- * (phones, most tablets) where `(hover: hover)` does not match.
- */
-export function usePrimaryInputCanHover(): boolean {
-  return useSyncExternalStore(subscribeHoverMedia, getHoverMediaSnapshot, getHoverMediaServerSnapshot)
-}
+import { createContext, useContext, type ReactNode } from 'react'
 
 export type SortableCompilationKind = 'collection' | 'playlist'
